+++ api/maven-api-model/src/main/mdo/maven.mdo
@@ -3315,50 +3315,24 @@
     /**
      * @see java.lang.Object#equals(java.lang.Object)
      */
-    public boolean equals( Object o )
-    {
-        if ( this == o )
-        {
+    public boolean equals(Object o) {
+        if (this == o) {
             return true;
-        }
-
-        if ( !( o instanceof Extension ) )
-        {
+        } else if (!(o instanceof Extension)) {
             return false;
+        } else {
+            Extension e = (Extension) o;
+            return java.util.Objects.equals(e.getArtifactId(), getArtifactId())
+                && java.util.Objects.equals(e.getGroupId(), getGroupId())
+                && java.util.Objects.equals(e.getVersion(), getVersion());
         }
-
-        Extension e = (Extension) o;
-
-        if ( !equal( e.getArtifactId(), getArtifactId() ) )
-        {
-            return false;
-        }
-        else if ( !equal( e.getGroupId(), getGroupId() ) )
-        {
-            return false;
-        }
-        else if ( !equal( e.getVersion(), getVersion() ) )
-        {
-            return false;
-        }
-        return true;
-    }
-
-    private static <T> boolean equal( T obj1, T obj2 )
-    {
-        return ( obj1 != null ) ? obj1.equals( obj2 ) : obj2 == null;
     }
 
     /**
      * @see java.lang.Object#hashCode()
      */
-    public int hashCode()
-    {
-        int result = 17;
-        result = 37 * result + ( getArtifactId() != null ? getArtifactId().hashCode() : 0 );
-        result = 37 * result + ( getGroupId() != null ? getGroupId().hashCode() : 0 );
-        result = 37 * result + ( getVersion() != null ? getVersion().hashCode() : 0 );
-        return result;
+    public int hashCode() {
+        return java.util.Objects.hash(getArtifactId(), getGroupId(), getVersion());
     }
             ]]>
           </code>
@@ -3367,7 +3341,7 @@
     </class>
     <class locationTracker="locations">
       <name>InputLocation</name>
-      <version>4.0.0+</version>
+      <version>4.0.0/4.0.99</version>
       <fields>
         <!-- line, column and source fields are auto-generated by Modello -->
       </fields>
@@ -3378,8 +3352,7 @@
             <![CDATA[
 
     @Override
-    public String toString()
-    {
+    public String toString() {
         return getLineNumber() + " : " + getColumnNumber() + ", " + getSource();
     }
             ]]>
@@ -3389,7 +3362,7 @@
     </class>
     <class sourceTracker="source">
       <name>InputSource</name>
-      <version>4.0.0+</version>
+      <version>4.0.0/4.0.99</version>
       <fields>
         <field>
           <name>modelId</name>
