+++ api/maven-api-plugin/src/main/mdo/lifecycle.mdo
@@ -18,20 +18,20 @@ under the License.
 -->
 
 <model xmlns="http://codehaus-plexus.github.io/MODELLO/2.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://codehaus-plexus.github.io/MODELLO/2.0.0 http://codehaus-plexus.github.io/modello/xsd/modello-2.0.0.xsd"
+  xsi:schemaLocation="http://codehaus-plexus.github.io/MODELLO/2.0.0 https://codehaus-plexus.github.io/modello/xsd/modello-2.0.0.xsd"
   xml.namespace="http://maven.apache.org/LIFECYCLE/${version}"
   xml.schemaLocation="http://maven.apache.org/xsd/lifecycle-${version}.xsd">
   <id>lifecycle</id>
   <name>Lifecycle</name>
   <description><![CDATA[
     Configuration of custom lifecycle mappings for the plugin, as generally stored in
-    <code>META-INF/maven/lifecycle.xml</code> in a plugin's jar artifact.
+    {@code META-INF/maven/lifecycle.xml} in a plugin's jar artifact.
   ]]></description>
   <classes>
     <class rootElement="true" xml.tagName="lifecycles" xsd.compositor="sequence">
       <name>LifecycleConfiguration</name>
       <version>1.0.0</version>
-      <description><![CDATA[Root element of the <code>lifecycle.xml</code> file.]]></description>
+      <description><![CDATA[Root element of the {@code lifecycle.xml} file.]]></description>
       <fields>
         <field>
           <name>lifecycles</name>
