+++ src/mdo/java/InputLocation.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.maven.api.settings;
+package ${package};
 
 import java.io.Serializable;
 import java.util.Collection;
